@@ -5,21 +5,23 @@ import com.legacy.monolith.entity.User;
 import com.legacy.monolith.messaging.MessagePublisher;
 import com.legacy.monolith.repository.OrderRepository;
 import com.legacy.monolith.repository.UserRepository;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class OrderServiceTest {
     
     @Mock
@@ -48,7 +50,7 @@ public class OrderServiceTest {
         order.setQuantity(2);
         order.setPrice(new BigDecimal("99.99"));
         
-        when(userRepository.findOne(1L)).thenReturn(user);
+        when(userRepository.findById(1L)).thenReturn(java.util.Optional.of(user));
         when(orderRepository.save(any(Order.class))).thenReturn(order);
         
         Order result = orderService.createOrder(order);
@@ -58,12 +60,14 @@ public class OrderServiceTest {
         verify(messagePublisher, times(1)).publishOrderCreated(any());
     }
     
-    @Test(expected = RuntimeException.class)
+    @Test
     public void testCreateOrderWithoutUser() {
+        assertThrows(RuntimeException.class, () -> {
         Order order = new Order();
         order.setProductName("Test Product");
         
         orderService.createOrder(order);
+        });
     }
     
     @Test
